Changes to classifier updated ModelData class moved BERT embedding code to its own file Added SVM training and validation
